Endicheck: introduce the empty tag and add shifts.
The 'EMPTY' tags are used to quickly identify empty bytes quickly. This
enables us to implement a very tolerant OR operation that allows the
user to merge different data, as long as the bytes are disjoint (see
bitpack.c test).

Also the 'EMPTY' tags will not be produced on arithmetic or similar
operations and this is actually a feature, we want them to appear only
when they have arisen from data shuffling operation.