Add support for guessing the size of a constant from value.
Often the compiler introduces a constant that is wider than what the C
type originally was and how it will be used. This might be due to
architectural constraints (no subregisters) or because it is more
efficient (x86).

So support setting the shadow value on actual constant value, mostly for
supporting the byte case. An example:
 shadow(0x000000AB) = N-N-N-A

If the constant is later narrowed into a byte, it will not trigger an
endianity error. But if we use the full bytes, it will still trigger an
error.