Always mark the first byte as any.
This should help lower false positive rate. In future, this could be
made into an option. A new test added and made to pass is casted_byte.c.

It also has drawbacks: error reporting is less precise, since it does
not mention the first byte. This should be fixed by a suitable
heuristic, however.