Change the origin tracking tests.
There is a slight problem with small functions, the instruction pointer
is incorrect at their beginning. This is not problem for memcheck, since
it can only fail at memory operations and produce otags only when stack
is moved.

To avoid this problem for now, let us just make the source() function
more complex. In the future, core fix should be made.